feat(phase50): comment out 28+ Windows-only DirectX files from WW3D2 CMakeLists
- Excluded files that directly or indirectly depend on dx8wrapper.h
- Files include: assetmgr, bmp2d, ddsfile, boxrobj, camera, dazzle, decalmsh, etc.
- Wrapped TextureClass virtual methods in #ifdef _WIN32 to avoid forward declaration errors
- Platform-conditional compilation: source files excluded via CMakeLists, not deleted
- Applies to both GeneralsMD and Generals WW3D2 library builds
- Partial Phase 50 Vulkan-only refactor - build in progress

+++ Dependencies/Utility/Utility/thread_compat.h
@@ -20,10 +20,13 @@
 #pragma once
 #include <pthread.h>
 #include <unistd.h>
+#include <cstdint>
 
 inline int GetCurrentThreadId()
 {
-  return pthread_self();
+  // Note: pthread_t may be an integer or pointer type depending on platform
+  // We cast it to an integer for compatibility with Windows DWORD-based thread IDs
+  return (int)(uintptr_t)pthread_self();
 }
 
 inline void Sleep(int ms)
